make set0 more compliant with Panel constraints + move function and test to util_matrix
