+++ include/dlaf/util_matrix.h
@@ -28,6 +28,7 @@ constexpr double M_PI = 3.141592;
 #include "dlaf/common/range2d.h"
 #include "dlaf/lapack/tile.h"
 #include "dlaf/matrix/matrix.h"
+#include "dlaf/matrix/panel.h"
 #include "dlaf/types.h"
 
 /// @file
@@ -139,6 +140,14 @@ template <class T, class ExecutorOrPolicy>
 void set0(ExecutorOrPolicy ex, Matrix<T, Device::CPU>& matrix) {
   for (const auto& idx : iterate_range2d(matrix.distribution().localNrTiles()))
     matrix(idx).then(ex, hpx::unwrapping(tile::set0<T>));
+}
+
+/// This functions sets all the elements of the tiles in the range to zero without inhibiting to set the
+/// tile as external afterward (e.g. the tile is set to zero without changing its state to "already used")
+template <class T, Coord axis, class ExecutorOrPolicy>
+void set0(ExecutorOrPolicy ex, Panel<axis, T, Device::CPU>& panel) {
+  for (const auto& tile_idx : panel.iteratorLocal())
+    panel(tile_idx).then(ex, hpx::unwrapping(tile::set0<T>));
 }
 
 /// Set the elements of the matrix.
