+++ test/unit/matrix/test_panel.cpp
@@ -394,45 +394,3 @@ TYPED_TEST(PanelTest, SetHeight) {
     checkPanelTileSize(default_dim, panel);
   }
 }
-
-template <class TypeParam, Coord panel_axis>
-void testSet0(const config_t& cfg, const comm::CommunicatorGrid& comm_grid) {
-  constexpr Coord coord1D = orthogonal(panel_axis);
-
-  Distribution dist(cfg.sz, cfg.blocksz, comm_grid.size(), comm_grid.rank(), {0, 0});
-  Panel<panel_axis, TypeParam, dlaf::Device::CPU> panel(dist, cfg.offset);
-
-  auto null_tile = [](const TileElementIndex&) { return TypeParam(0); };
-
-  for (SizeType head = cfg.offset.get<coord1D>(), tail = dist.nrTiles().get(coord1D); head <= tail;
-       ++head, --tail) {
-    panel.setRange(GlobalTileIndex(coord1D, head), GlobalTileIndex(coord1D, tail));
-
-    for (const auto& idx : panel.iteratorLocal())
-      hpx::dataflow(hpx::unwrapping(tile::laset<TypeParam>), lapack::MatrixType::General, 1, 1,
-                    panel(idx));
-
-    set0(hpx::launch::sync, panel);
-
-    // Note:
-    // `set0` does not inhibit from setting a tile as external.
-    //
-    // Unfortunately, this can be tested, because as soon as I access the panel
-    // tile (either RW or RO) for checking, I cannot test anymore that I can
-    // link that specific tile to an external one with `setTile`.
-
-    for (const auto& idx : panel.iteratorLocal())
-      CHECK_TILE_EQ(null_tile, panel.read(idx).get());
-
-    panel.reset();
-  }
-}
-
-TYPED_TEST(PanelTest, Set0) {
-  for (auto& comm_grid : this->commGrids()) {
-    for (const auto& cfg : test_params) {
-      testSet0<TypeParam, Coord::Col>(cfg, comm_grid);
-      testSet0<TypeParam, Coord::Row>(cfg, comm_grid);
-    }
-  }
-}
